+++ src/components/Location.tsx
@@ -8,38 +8,48 @@ interface LocationProps {
 }
 
 const Location = ({ name, addr, services }: LocationProps) => {
-  const [show, setShow] = useState(true);
-  const click = function () {
-    console.log(show);
-    setShow(!show);
+  const [hide, setHide] = useState(true);
+  const click = () => {
+    setHide(!hide);
   };
+  const rowOneStyle =
+    "md:pt-2 xl:pb-4 xl:pt-12 border-b-2 border-solid border-black xl:border-b-4";
+
+  const rowTwoStyle = "pb-2 pt-2 sm:pb-3 sm:pt-3 xl:pt-8";
   return (
-    <div onClick={click}>
-      <button
-        className="border-b-solid mt-1 flex w-full items-center border-b-2 border-black text-xs font-bold text-black md:mt-0 lg:mt-3 lg:border-b-4 lg:text-xl xl:mt-5 xl:pb-2 xl:text-[33px]"
-        onClick={click}
-      >
-        <div className="grid grid-cols-2 justify-items-start">
-          <div>FGA CYC {name}</div>
-          <div className="w-5 justify-self-end lg:w-7 xl:w-auto">
+    <>
+      <tr onClick={click}>
+        <th className={rowOneStyle}>
+          <div className="flex items-start text-xs font-bold sm:text-sm md:text-[18px] lg:text-xl xl:text-[33px]">
+            {"FGA CYC " + name}
+          </div>
+        </th>
+        <th className={rowOneStyle}>
+          <div
+            className={
+              hide
+                ? "w-[35px] lg:w-10 xl:w-14 2xl:w-16"
+                : "w-[24px] lg:w-8 xl:w-12 2xl:w-16"
+            }
+          >
             <Image
-              src={"/icons/right_arrow.svg"}
-              width={42}
-              height={42}
+              src={hide ? "/icons/right_arrow.svg" : "/icons/cross.svg"}
+              width={1}
+              height={1}
               alt="right arrow icon"
-              className="ml-3"
+              className="w-full pl-3"
             />
           </div>
-        </div>
-      </button>
-      <div
-        className={
-          (show ? "hidden" : "") +
-          " mt-3 flex text-[8px] font-bold sm:text-[10px] md:text-xs lg:text-sm xl:mt-4 xl:text-xl"
-        }
+        </th>
+      </tr>
+      <tr
+        hidden={hide}
+        className="text-[8px] font-bold sm:text-[10px] md:text-xs lg:text-sm xl:text-xl"
       >
-        <div className="w-[25vw] xl:w-[367px]">{addr}</div>
-        <div className={(show ? "hidden " : "") + " ms-2"}>
+        <th colSpan={2} className={"w-0 " + rowTwoStyle}>
+          <div className="mr-3 text-left">{addr}</div>
+        </th>
+        <th className={"text-left " + rowTwoStyle}>
           {services.map((s) => {
             return (
               <>
@@ -48,9 +58,9 @@ const Location = ({ name, addr, services }: LocationProps) => {
             );
           })}
           <div className="text-[#00edc2]">Get Directions {">"}</div>
-        </div>
-      </div>
-    </div>
+        </th>
+      </tr>
+    </>
   );
 };
